refactor: split user config commands by concern
