+++ internal/menu/user_config.go
@@ -1,23 +1,17 @@
 package menu
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"io"
 	"log/slog"
-	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/gliderlabs/ssh"
-	"golang.org/x/crypto/bcrypt"
 	"golang.org/x/term"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/ansi"
-	"github.com/ViSiON-3/vision-3-bbs/internal/editor"
 	"github.com/ViSiON-3/vision-3-bbs/internal/terminalio"
 	"github.com/ViSiON-3/vision-3-bbs/internal/user"
 )
@@ -64,180 +58,6 @@ func runCfgToggle(
 		stateStr = e.LoadedStrings.CfgToggleOn
 	}
 	msg := fmt.Sprintf(e.LoadedStrings.CfgToggleFormat, fieldName, stateStr)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(500 * time.Millisecond)
-	return currentUser, "", nil
-}
-
-func runCfgHotKeys(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	sessionStartTime := c.sessionStartTime
-	outputMode := c.outputMode
-
-	return runCfgToggle(e, s, terminal, userManager, currentUser, nodeNumber, sessionStartTime, args, outputMode,
-		"Hot Keys",
-		func(u *user.User) bool { return u.HotKeys },
-		func(u *user.User, v bool) { u.HotKeys = v },
-	)
-}
-
-func runCfgMorePrompts(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	sessionStartTime := c.sessionStartTime
-	outputMode := c.outputMode
-
-	return runCfgToggle(e, s, terminal, userManager, currentUser, nodeNumber, sessionStartTime, args, outputMode,
-		"More Prompts",
-		func(u *user.User) bool { return u.MorePrompts },
-		func(u *user.User, v bool) { u.MorePrompts = v },
-	)
-}
-
-func runCfgScreenWidth(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	current := currentUser.ScreenWidth
-	if current == 0 {
-		current = 80
-	}
-	prompt := fmt.Sprintf(e.LoadedStrings.CfgScreenWidthPrompt, current)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(prompt)), outputMode)
-
-	input, err := readLineFromSessionIH(s, terminal)
-	if err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-		return currentUser, "", nil
-	}
-
-	input = strings.TrimSpace(input)
-	if input == "" {
-		return currentUser, "", nil
-	}
-
-	val, parseErr := strconv.Atoi(input)
-	if parseErr != nil || val < 40 || val > 255 {
-		msg := e.LoadedStrings.CfgScreenWidthInvalid
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-		time.Sleep(500 * time.Millisecond)
-		return currentUser, "", nil
-	}
-
-	currentUser.ScreenWidth = val
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		slog.Error("failed to save screen width", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg := fmt.Sprintf(e.LoadedStrings.CfgScreenWidthSet, val)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(500 * time.Millisecond)
-	return currentUser, "", nil
-}
-
-func runCfgScreenHeight(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	current := currentUser.ScreenHeight
-	if current == 0 {
-		current = 25
-	}
-	prompt := fmt.Sprintf(e.LoadedStrings.CfgScreenHeightPrompt, current)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(prompt)), outputMode)
-
-	input, err := readLineFromSessionIH(s, terminal)
-	if err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-		return currentUser, "", nil
-	}
-
-	input = strings.TrimSpace(input)
-	if input == "" {
-		return currentUser, "", nil
-	}
-
-	val, parseErr := strconv.Atoi(input)
-	if parseErr != nil || val < 21 || val > 60 {
-		msg := e.LoadedStrings.CfgScreenHeightInvalid
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-		time.Sleep(500 * time.Millisecond)
-		return currentUser, "", nil
-	}
-
-	currentUser.ScreenHeight = val
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		slog.Error("failed to save screen height", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg := fmt.Sprintf(e.LoadedStrings.CfgScreenHeightSet, val)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(500 * time.Millisecond)
-	return currentUser, "", nil
-}
-
-func runCfgTermType(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	current := currentUser.OutputMode
-	if current == "" {
-		current = "cp437"
-	}
-
-	if current == "ansi" {
-		currentUser.OutputMode = "cp437"
-	} else {
-		currentUser.OutputMode = "ansi"
-	}
-
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		slog.Error("failed to save terminal type", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg := fmt.Sprintf(e.LoadedStrings.CfgTermTypeSet, strings.ToUpper(currentUser.OutputMode))
 	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
 	time.Sleep(500 * time.Millisecond)
 	return currentUser, "", nil
@@ -290,541 +110,4 @@ func runCfgStringInput(
 	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
 	time.Sleep(500 * time.Millisecond)
 	return currentUser, "", nil
-}
-
-func runCfgRealName(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	return runCfgStringInput(e, s, terminal, userManager, currentUser, nodeNumber, outputMode,
-		"Real Name", 40,
-		func(u *user.User) string { return u.RealName },
-		func(u *user.User, v string) { u.RealName = v },
-	)
-}
-
-func runCfgNote(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	return runCfgStringInput(e, s, terminal, userManager, currentUser, nodeNumber, outputMode,
-		"User Note", 35,
-		func(u *user.User) string { return u.PrivateNote },
-		func(u *user.User, v string) { u.PrivateNote = v },
-	)
-}
-
-func runCfgPassword(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-	termWidth := c.termWidth
-	termHeight := c.termHeight
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	// Prompt for current password
-	msg := e.LoadedStrings.CfgCurrentPwPrompt
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-
-	oldPw, err := readPasswordSecurely(s, terminal, outputMode)
-	if err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-		return currentUser, "", nil
-	}
-
-	// Verify current password
-	if bcryptErr := bcrypt.CompareHashAndPassword([]byte(currentUser.PasswordHash), []byte(oldPw)); bcryptErr != nil {
-		msg := e.LoadedStrings.CfgIncorrectPw
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-		time.Sleep(1 * time.Second)
-		return currentUser, "", nil
-	}
-
-	// Prompt for new password using existing helper
-	newPw, err := e.promptForPassword(s, terminal, nodeNumber, outputMode, termWidth, termHeight)
-	if err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-		return currentUser, "", nil
-	}
-	if newPw == "" {
-		return currentUser, "", nil
-	}
-
-	// Hash and save
-	hashed, err := bcrypt.GenerateFromPassword([]byte(newPw), bcrypt.DefaultCost)
-	if err != nil {
-		slog.Error("failed to hash new password", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	currentUser.PasswordHash = string(hashed)
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		slog.Error("failed to save password", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg = e.LoadedStrings.CfgPasswordChanged
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(1 * time.Second)
-	return currentUser, "", nil
-}
-
-var colorSlotNames = [7]string{"Prompt", "Input", "Text", "Stat", "Text2", "Stat2", "Bar"}
-
-func runCfgColor(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	slot := 0
-	if trimmed := strings.TrimSpace(args); trimmed != "" {
-		if parsed, err := strconv.Atoi(trimmed); err == nil && parsed >= 0 && parsed < 7 {
-			slot = parsed
-		}
-	}
-
-	slotName := colorSlotNames[slot]
-
-	// Display color palette
-	var palette strings.Builder
-	fmt.Fprintf(&palette, e.LoadedStrings.CfgColorSelectPrompt, slotName)
-	for i := 0; i < 16; i++ {
-		fmt.Fprintf(&palette, "|%02d  %2d  ", i, i)
-		if i == 7 {
-			palette.WriteString("\r\n")
-		}
-	}
-	palette.WriteString(e.LoadedStrings.CfgColorInputPrompt)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(palette.String())), outputMode)
-
-	input, err := readLineFromSessionIH(s, terminal)
-	if err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-		return currentUser, "", nil
-	}
-
-	input = strings.TrimSpace(input)
-	if input == "" {
-		return currentUser, "", nil
-	}
-
-	val, parseErr := strconv.Atoi(input)
-	if parseErr != nil || val < 0 || val > 15 {
-		msg := e.LoadedStrings.CfgColorInvalid
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-		time.Sleep(500 * time.Millisecond)
-		return currentUser, "", nil
-	}
-
-	currentUser.Colors[slot] = val
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		slog.Error("failed to save color", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg := fmt.Sprintf(e.LoadedStrings.CfgColorSet, slotName, val, val)
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(500 * time.Millisecond)
-	return currentUser, "", nil
-}
-
-func runCfgViewConfig(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-	termWidth := c.termWidth
-	termHeight := c.termHeight
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	topPath := filepath.Join(e.MenuSetPath, "templates", "USRCFGV.TOP")
-	botPath := filepath.Join(e.MenuSetPath, "templates", "USRCFGV.BOT")
-
-	topBytes, err := os.ReadFile(topPath)
-	if err != nil && !os.IsNotExist(err) {
-		slog.Warn("failed to read template", "node", nodeNumber, "path", topPath, "error", err)
-	}
-	botBytes, err := os.ReadFile(botPath)
-	if err != nil && !os.IsNotExist(err) {
-		slog.Warn("failed to read template", "node", nodeNumber, "path", botPath, "error", err)
-	}
-
-	topBytes = stripSauceMetadata(topBytes)
-	botBytes = stripSauceMetadata(botBytes)
-	topBytes = normalizePipeCodeDelimiters(topBytes)
-	botBytes = normalizePipeCodeDelimiters(botBytes)
-
-	var buf bytes.Buffer
-
-	if len(topBytes) > 0 {
-		buf.Write(ansi.ReplacePipeCodes(topBytes))
-		buf.WriteString("\r\n")
-	}
-
-	boolStr := func(v bool) string {
-		if v {
-			return e.LoadedStrings.CfgToggleOn
-		}
-		return e.LoadedStrings.CfgToggleOff
-	}
-
-	width := currentUser.ScreenWidth
-	if width == 0 {
-		width = 80
-	}
-	height := currentUser.ScreenHeight
-	if height == 0 {
-		height = 25
-	}
-	outMode := currentUser.OutputMode
-	if outMode == "" {
-		outMode = "cp437"
-	}
-
-	lines := []string{
-		fmt.Sprintf(e.LoadedStrings.CfgViewScreenWidth, width),
-		fmt.Sprintf(e.LoadedStrings.CfgViewScreenHeight, height),
-		fmt.Sprintf(e.LoadedStrings.CfgViewTermType, strings.ToUpper(outMode)),
-		fmt.Sprintf(e.LoadedStrings.CfgViewHotKeys, boolStr(currentUser.HotKeys)),
-		fmt.Sprintf(e.LoadedStrings.CfgViewMorePrompts, boolStr(currentUser.MorePrompts)),
-		fmt.Sprintf(e.LoadedStrings.CfgViewFileListMode, fileListModeDisplay(currentUser.FileListingMode)),
-		fmt.Sprintf(e.LoadedStrings.CfgViewMsgHeader, currentUser.MsgHdr),
-		fmt.Sprintf(e.LoadedStrings.CfgViewCustomPrompt, currentUser.CustomPrompt),
-		"",
-		fmt.Sprintf(e.LoadedStrings.CfgViewPromptColor, currentUser.Colors[0], currentUser.Colors[0], currentUser.Colors[1], currentUser.Colors[1]),
-		fmt.Sprintf(e.LoadedStrings.CfgViewTextColor, currentUser.Colors[2], currentUser.Colors[2], currentUser.Colors[3], currentUser.Colors[3]),
-		fmt.Sprintf(e.LoadedStrings.CfgViewText2Color, currentUser.Colors[4], currentUser.Colors[4], currentUser.Colors[5], currentUser.Colors[5]),
-		fmt.Sprintf(e.LoadedStrings.CfgViewBarColor, currentUser.Colors[6], currentUser.Colors[6]),
-		"",
-		fmt.Sprintf(e.LoadedStrings.CfgViewRealName, currentUser.RealName),
-		fmt.Sprintf(e.LoadedStrings.CfgViewNote, currentUser.PrivateNote),
-	}
-
-	// Append auto-signature info
-	if currentUser.AutoSignature != "" {
-		lines = append(lines, "", "|03  Auto-Signature:|07")
-		sigLines := strings.Split(currentUser.AutoSignature, "\n")
-		for _, sl := range sigLines {
-			lines = append(lines, "    "+sl)
-		}
-	} else {
-		lines = append(lines, "", "|08  Auto-Signature: (none)|07")
-	}
-
-	for _, line := range lines {
-		buf.Write(ansi.ReplacePipeCodes([]byte(line)))
-		buf.WriteString("\r\n")
-	}
-
-	if len(botBytes) > 0 {
-		buf.Write(ansi.ReplacePipeCodes(botBytes))
-		buf.WriteString("\r\n")
-	}
-
-	terminalio.WriteProcessedBytes(terminal, buf.Bytes(), outputMode)
-
-	// Pause
-	pausePrompt := e.LoadedStrings.PauseString
-	if pausePrompt == "" {
-		pausePrompt = "\r\n|07Press |15[ENTER]|07 to continue... "
-	}
-	if err := writeCenteredPausePrompt(s, terminal, pausePrompt, outputMode, termWidth, termHeight); err != nil {
-		if errors.Is(err, io.EOF) {
-			return nil, "LOGOFF", io.EOF
-		}
-	}
-
-	return currentUser, "", nil
-}
-
-func runCfgFileListMode(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	originalMode := currentUser.FileListingMode
-	if strings.ToLower(originalMode) != "classic" {
-		currentUser.FileListingMode = "classic"
-	} else {
-		currentUser.FileListingMode = "lightbar"
-	}
-
-	if err := userManager.UpdateUser(currentUser); err != nil {
-		currentUser.FileListingMode = originalMode
-		slog.Error("failed to save file listing mode", "node", nodeNumber, "error", err)
-		return currentUser, "", nil
-	}
-
-	msg := fmt.Sprintf(e.LoadedStrings.CfgFileListModeSet, fileListModeDisplay(currentUser.FileListingMode))
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(msg)), outputMode)
-	time.Sleep(500 * time.Millisecond)
-	return currentUser, "", nil
-}
-
-// maxAutoSigLines is the maximum number of lines allowed in an auto-signature.
-const maxAutoSigLines = 5
-
-func runCfgAutoSig(c *cmdCtx, args string) (*user.User, string, error) {
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	for {
-		// Display header
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|15Auto-Signature|07\r\n")), outputMode)
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("|08An Auto-Signature is appended to the end of any message you post.|07\r\n\r\n")), outputMode)
-
-		// Show current signature
-		if currentUser.AutoSignature == "" {
-			terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("|03You currently do not have an Auto-Signature.|07\r\n\r\n")), outputMode)
-		} else {
-			terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("|03Your current Auto-Signature is...|07\r\n\r\n")), outputMode)
-			sigLines := strings.Split(currentUser.AutoSignature, "\n")
-			for _, line := range sigLines {
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(line)), outputMode)
-				terminalio.WriteProcessedBytes(terminal, []byte("\r\n"), outputMode)
-			}
-			terminalio.WriteProcessedBytes(terminal, []byte("\r\n"), outputMode)
-		}
-
-		// Menu prompt
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("|09C|07hange/create  |09D|07elete  |09Q|07uit : ")), outputMode)
-
-		input, err := readLineFromSessionIH(s, terminal)
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				return nil, "LOGOFF", io.EOF
-			}
-			return currentUser, "", nil
-		}
-
-		input = strings.TrimSpace(strings.ToUpper(input))
-		if input == "" || input == "Q" {
-			return currentUser, "", nil
-		}
-
-		switch input {
-		case "C":
-			// Launch the ANSI editor with current signature as initial content
-			ih := getSessionIH(s)
-			terminalio.WriteProcessedBytes(terminal, []byte(ansi.ClearScreen()), outputMode)
-			editorCtx := editor.EditorContext{
-				NodeNumber: nodeNumber,
-				ConfArea:   "Auto-Signature",
-			}
-			body, saved, edErr := editor.RunEditorWithMetadata(
-				currentUser.AutoSignature, s, s, outputMode,
-				"Auto-Signature", "All", currentUser.Handle, false,
-				"", "", "", "", false, nil, ih, editorCtx,
-			)
-			terminalio.WriteProcessedBytes(terminal, []byte(ansi.ClearScreen()), outputMode)
-
-			if edErr != nil {
-				slog.Error("editor failed for auto-sig", "node", nodeNumber, "error", edErr)
-				return currentUser, "", nil
-			}
-			if !saved {
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|07Auto-Signature not changed.\r\n")), outputMode)
-				time.Sleep(500 * time.Millisecond)
-				continue
-			}
-
-			// Truncate to maxAutoSigLines lines
-			body = strings.TrimRight(body, "\r\n")
-			if body != "" {
-				lines := strings.Split(body, "\n")
-				if len(lines) > maxAutoSigLines {
-					lines = lines[:maxAutoSigLines]
-					terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(
-						fmt.Sprintf("\r\n|03Signature truncated to %d lines.|07\r\n", maxAutoSigLines),
-					)), outputMode)
-				}
-				body = strings.Join(lines, "\n")
-			}
-
-			currentUser.AutoSignature = body
-			if err := userManager.UpdateUser(currentUser); err != nil {
-				slog.Error("failed to save auto-signature", "node", nodeNumber, "error", err)
-				return currentUser, "", nil
-			}
-			if body == "" {
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|03Auto-Signature cleared.|07\r\n")), outputMode)
-			} else {
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|02Auto-Signature saved!|07\r\n")), outputMode)
-			}
-			time.Sleep(500 * time.Millisecond)
-
-		case "D":
-			if currentUser.AutoSignature == "" {
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|03You don't have an Auto-Signature to delete!|07\r\n")), outputMode)
-			} else {
-				currentUser.AutoSignature = ""
-				if err := userManager.UpdateUser(currentUser); err != nil {
-					slog.Error("failed to delete auto-signature", "node", nodeNumber, "error", err)
-					return currentUser, "", nil
-				}
-				terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte("\r\n|03Auto-Signature has been deleted.|07\r\n")), outputMode)
-			}
-			time.Sleep(500 * time.Millisecond)
-		}
-	}
-}
-
-func runCfgFileColumns(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	boolStr := func(v bool) string {
-		if v {
-			return e.LoadedStrings.CfgToggleOn
-		}
-		return e.LoadedStrings.CfgToggleOff
-	}
-
-	for {
-		c := currentUser.FileListColumns
-		allDefault := !c.Name && !c.Size && !c.Date && !c.Downloads && !c.Uploader && !c.Description
-
-		displayState := func(val bool) string {
-			if allDefault {
-				return boolStr(true)
-			}
-			return boolStr(val)
-		}
-
-		var buf strings.Builder
-		buf.WriteString(e.LoadedStrings.CfgFileColumnsHeader)
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "N", "Name", displayState(c.Name))
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "S", "Size", displayState(c.Size))
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "D", "Date", displayState(c.Date))
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "L", "Downloads", displayState(c.Downloads))
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "U", "Uploader", displayState(c.Uploader))
-		fmt.Fprintf(&buf, e.LoadedStrings.CfgFileColumnsToggle, "E", "Description", displayState(c.Description))
-		buf.WriteString(e.LoadedStrings.CfgFileColumnsHeader) // reuse as prompt separator
-		terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(buf.String())), outputMode)
-
-		input, err := readLineFromSessionIH(s, terminal)
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				return nil, "LOGOFF", io.EOF
-			}
-			return currentUser, "", nil
-		}
-
-		input = strings.TrimSpace(strings.ToUpper(input))
-		if input == "" || input == "Q" {
-			if err := userManager.UpdateUser(currentUser); err != nil {
-				slog.Error("failed to save file column preferences", "node", nodeNumber, "error", err)
-			}
-			terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(e.LoadedStrings.CfgFileColumnsSaved)), outputMode)
-			time.Sleep(500 * time.Millisecond)
-			return currentUser, "", nil
-		}
-
-		if allDefault {
-			currentUser.FileListColumns.Name = true
-			currentUser.FileListColumns.Size = true
-			currentUser.FileListColumns.Date = true
-			currentUser.FileListColumns.Downloads = true
-			currentUser.FileListColumns.Uploader = true
-			currentUser.FileListColumns.Description = true
-		}
-
-		switch input {
-		case "N":
-			currentUser.FileListColumns.Name = !currentUser.FileListColumns.Name
-		case "S":
-			currentUser.FileListColumns.Size = !currentUser.FileListColumns.Size
-		case "D":
-			currentUser.FileListColumns.Date = !currentUser.FileListColumns.Date
-		case "L":
-			currentUser.FileListColumns.Downloads = !currentUser.FileListColumns.Downloads
-		case "U":
-			currentUser.FileListColumns.Uploader = !currentUser.FileListColumns.Uploader
-		case "E":
-			currentUser.FileListColumns.Description = !currentUser.FileListColumns.Description
-		}
-	}
-}
-
-func runCfgCustomPrompt(c *cmdCtx, args string) (*user.User, string, error) {
-	e := c.e
-	s := c.s
-	terminal := c.terminal
-	userManager := c.userManager
-	currentUser := c.currentUser
-	nodeNumber := c.nodeNumber
-	outputMode := c.outputMode
-
-	if currentUser == nil {
-		return nil, "", nil
-	}
-
-	help := e.LoadedStrings.CfgCustomPromptHelp
-	terminalio.WriteProcessedBytes(terminal, ansi.ReplacePipeCodes([]byte(help)), outputMode)
-
-	return runCfgStringInput(e, s, terminal, userManager, currentUser, nodeNumber, outputMode,
-		"Custom Prompt", 80,
-		func(u *user.User) string { return u.CustomPrompt },
-		func(u *user.User, v string) { u.CustomPrompt = v },
-	)
 }
